refactor(admin): move the dashboard script out of html.rs
Adding the grouped-accounts renderer pushed src/admin/html.rs to 581
lines, past the contribution checklist's 500-line guidance.

Split the inline dashboard script into src/admin/script.rs: html.rs
drops to 262 lines and script.rs is 330. Holding it as a plain const
rather than a format! argument also lets the JavaScript read as
JavaScript — inside a format string every brace has to be doubled — so
the single {csrf} placeholder is substituted with replace() instead.

Behaviour-preserving: dashboard_page() renders byte-for-byte identical
output before and after (33348 bytes, verified by diffing the rendered
page across the change).
